+++ src/select/index.jsx
@@ -38,6 +38,7 @@ const paginationButtonStyles = {}
  * @property {function(): import("react").ReactNode} [currentContent]
  * @property {boolean} [disabled]
  * @property {string} [html]
+ * @property {import("react").ReactNode} [right]
  *
  * Additional option props are allowed; HayaSelect only uses the keys above.
  */
@@ -188,6 +189,7 @@ export default memo(shapeComponent(class HayaSelect extends ShapeComponent {
         currentContent: PropTypes.func,
         disabled: PropTypes.bool,
         html: PropTypes.string,
+        right: PropTypes.node,
         text: PropTypes.node,
         value: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired
       })),
@@ -1361,6 +1363,7 @@ export default memo(shapeComponent(class HayaSelect extends ShapeComponent {
     const toggleOption = toggleOptions?.find((toggleOptionI) => toggleOptionI.value == toggleValue)
     const selected = this.getCurrentOptionValues().some((value) => value == option.value)
     let style
+    let contentNode
 
     if (mode == "current") {
       style = this.stylingFor("currentOptionPresentationText", {flex: 1, whiteSpace: "nowrap"})
@@ -1376,7 +1379,7 @@ export default memo(shapeComponent(class HayaSelect extends ShapeComponent {
           toggleValue: toggleOption?.value,
           value: option.value
         }, [option.text, option.value, toggleOption?.icon, toggleOption?.value])}
-        style={this.cache("optionPresentationStyle", {flexDirection: "row"})}
+        style={this.cache("optionPresentationStyle", {flexDirection: "row", alignItems: "center"})}
         testID="option-presentation"
       >
         {toggleOptions && !(option.value in toggled) &&
@@ -1395,17 +1398,17 @@ export default memo(shapeComponent(class HayaSelect extends ShapeComponent {
         }
         {(() => {
           if (optionContent) {
-            return optionContent({icon, mode, option, selected, toggleOption, toggleValue, toggled})
+            contentNode = optionContent({icon, mode, option, selected, toggleOption, toggleValue, toggled})
           } else if (mode == "current" && option.currentContent) {
-            return option.currentContent()
+            contentNode = option.currentContent()
           } else if (option.content) {
-            return option.content()
+            contentNode = option.content()
           } else if ("html" in option && Platform.OS != "web") {
-            return <RenderHtml source={{html: digg(option, "html")}} />
+            contentNode = <RenderHtml source={{html: digg(option, "html")}} />
           } else if ("html" in option && Platform.OS == "web") {
-            return <div dangerouslySetInnerHTML={{__html: digg(option, "html")}} />
+            contentNode = <div dangerouslySetInnerHTML={{__html: digg(option, "html")}} />
           } else {
-            return (
+            contentNode = (
               <Text
                 style={style}
                 testID="option-presentation-text"
@@ -1414,6 +1417,19 @@ export default memo(shapeComponent(class HayaSelect extends ShapeComponent {
               </Text>
             )
           }
+
+          return (
+            <>
+              <View style={this.cache("optionPresentationContentStyle", {flex: 1})}>
+                {contentNode}
+              </View>
+              {option.right &&
+                <View style={this.cache("optionPresentationRightStyle", {alignItems: "center", justifyContent: "center", marginLeft: 8})}>
+                  {option.right}
+                </View>
+              }
+            </>
+          )
         })()}
       </View>
     )